Add a `--no-color` flag
Added a `--no-color` flag that disables pterm coloring for the following commands that use table output:

- `profile list`
- `stack list`
- `stack show`

The reason I added it to individual commands rather than adding the option to the root command is because of the way the urfave cli library works - that would have caused us to have to write commands like `spacectl --no-color stack list` instead of `spacectl stack list --no-color`.

In addition, if we detect the output is not a terminal, we disable coloring automatically.

Issues: #52