Add run list command
Added a new command to list the runs for a stack.

I've tried to keep the amount of data in the table output minimal while still being useful, but we'll probably need to tweak it in future. The JSON output only contains fields that are relevant for a tracked run.

Issues: #75