Use filepath functions for path manipulation in getGitRepositorySubdir
Before this commit, the root directory would be removed which leaves a
trailing file separator character in the beginning. This would then be trimmed
by using strings.TrimPrefix. The problem here is that it uses a hardcoded
file separator character which won't work on Windows because
Windows uses a backslash instead of a forward slash.

Using filepath.Rel to get the directory relative to the root directory
will not leave a separator character in the beginning.

It's also important to use filepath.ToSlash which will translate every
file separator character to the forward slash.

This commit makes discovery work on Windows which it didn't before.