Do not recurse into ignored directories.
When generating the tar.gz for local previews can skip recursively
processing them. This potentially saves a lot of processing time by
removing the stat calls to the files contained in them. This tends to be
the case for node_modules and other commonly ignored paths.

On my internal git repo running on macos this results in a -82% change
into the to generate the tar.tz.
```
$ time spacectl stack local-preview --id test  --no-upload
151.33 secs

$ time ~/src/spacectl/spacectl stack local-preview --id test  --no-upload
26.32 secs
```

I also checked if the same files were included in both tar.gz files.
```
$ tar -tf ~/old.tar.gz | sed 's|[^/]*||' | shasum
82510ad099d3eb9e3c1bba906b8b8dff00c802aa  -

$ tar -tf ~/new.tar.gz | sed 's|[^/]*||' | shasum
82510ad099d3eb9e3c1bba906b8b8dff00c802aa  -
```