feat: add support for versioning spacectl commands
I've created a new concept of there being multiple versions of each command. For each command, we specify an earliest supported version, which can be one of:

- `all` - supported on any Spacelift version, both SaaS and Self-Hosted.
- `latest` - only supported by SaaS. It will be supported by Self-Hosted after the next release.
- A specific version number, e.g. `2.5.0`.

The idea is that on startup, we can query for the current version of the Spacelift instance, and then filter by commands that are supported. If we have to add a new command or update an existing command in a way that relies on new changes to the GraphQL API, we can add a new command version instead to maintain backwards compatibility.

When a new Self-Hosted version is released, we can run a script to automatically find any commands which are set to `cmd.SupportedVersionLatest` and replace them with the actual version number.

Right now this is just a WIP to illustrate the idea, and the following is missing:

- Wire up querying for the version before building the commands. We'll want to add logic to avoid doing this for certain commands like `profile login` and `version`, and make it best effort for commands like `help` (or maybe display a notice explaining that the full available commands will be listed after connecting to a Spacelift instance).
- Handle loading subcommands recursively. Right now I just hacked something together quickly that only works for a single level of the tree.
- Migrate all commands over to the new approach.