feat: make GraphQL request to determine Spacelift version
I've wired up the code to actually get the version of the Spacelift instance that spacectl is talking to on startup. If we can't figure out the version for some reason (e.g. the user hasn't logged in, their credentials have expired, or they're talking to a very old version that doesn't support the debugInfo endpoint), the code treats the version as unknown and only returns commands known to work against all versions.