fix: ignore prerelease part of Self-Hosted version numbers
I've updated the code that gets the version from the API to ignore prerelease or metadata info in the version number. This is because we typically use versions like `v3.0.0-alpha.1` in preparation for a Self-Hosted release, but SemVer considers that to be an older version than `v3.0.0` and so we wouldn't show commands compatible with a v3.0.0 instance.