fix: update `provider create-version` to use versioned commands (#309)
The command was already trying to be backwards-compatible because we were parsing the GraphQL schema to check if the mutation was available, but unfortunately that wasn't quite enough because it didn't take into account that we can't query for fields that don't exist on older versions.

The new `terraformProviderVersionRegisterPlatformV2` command was introduced in Self-Hosted version 2.3.0, and I think the new upload headers were introduced in 2.5.0, but I've actually just said that the new version should be used for v3.0.0 and above. The reason is that it's only needed for v3+ for Spacelift instances hosted outside of AWS.